pythonrag/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv

pypdf

langchain
langchain-text-splitters
langchain-community
langchain-google-genai
google-generativeai

psycopg2-binary
pgvector
